refactor!: move URL helpers to `@crawlee/core`
